add: prettierrc 설정
