+++ .prettierrc
@@ -1,0 +1,7 @@
+{
+  "printWidth": 80,
+  "tabWidth": 2,
+  "singleQuote": false,
+  "trailingComma": "all",
+  "semi": false
+}
