+++ src/components/Header.jsx
@@ -1,0 +1,9 @@
+function Header() {
+    return (
+        <header>
+            <strong>TO DO LIST</strong>
+        </header>
+    )
+}
+
+export default Header;
